Updated Readme.MD, consolidated example components into custom-thumbnails.tsx
