+++ components/Markdown.tsx
@@ -3,8 +3,6 @@ import unified from "unified";
 import parse from "remark-parse";
 import remark2react from "remark-react";
 
-import { media } from "../styles/mixins";
-
 type MarkdownWrapperProps = {
   markdown?: string;
   styleOverrides?: any;
@@ -19,6 +17,9 @@ export default function MarkdownWrapper({
       css={[
         css`
           width: 100%;
+          * {
+            color: var(--black);
+          }
           h2,
           h3,
           h4 {
@@ -38,9 +39,19 @@ export default function MarkdownWrapper({
             text-decoration: underline;
             display: inline-block;
           }
+          ol,
+          ul {
+            margin-block-start: 10px;
+            margin-block-end: 10px;
+            margin-inline-start: 0px;
+            margin-inline-end: 0px;
+            padding-inline-start: 40px;
+          }
           ol {
-            padding-left: var(--space-sm);
             list-style: number;
+          }
+          ul {
+            list-style-type: disc;
           }
           pre {
             font-size: var(--text-01) !important;
@@ -53,10 +64,14 @@ export default function MarkdownWrapper({
             color: var(--black) !important;
             overflow-x: scroll;
             position: relative;
-            code {
-              font-family: SFMono-Regular, Consolas, Liberation Mono, Menlo,
+          }
+          code {
+            font-size: var(--text-01) !important;
+            line-height: 1.45;
+            border-radius: 5px;
+            background-color: var(--bg-color);
+            font-family: SFMono-Regular, Consolas, Liberation Mono, Menlo,
                 monospace !important;
-            }
           }
         `,
         styleOverrides,
